+++ relay-event-normalization/src/normalize/span/tag_extraction.rs
@@ -333,10 +333,17 @@ pub fn extract_tags(
                     .and_then(|value| value.as_str())
                 {
                     let lowercase_host = server_host.to_lowercase();
-                    let mut parts = lowercase_host.split(':');
+                    let (domain, port) = match lowercase_host.split_once(':') {
+                        Some((domain, port_str)) => {
+                            if let Ok(port) = port_str.parse::<u16>() {
+                                (domain, Some(port))
+                            } else {
+                                (domain, None)
+                            }
+                        }
+                        None => (server_host, None),
+                    };
 
-                    let domain = parts.next().unwrap_or("");
-                    let port = parts.next().and_then(|s| s.parse::<u16>().ok());
                     if let Some(url_scheme) = span
                         .data
                         .value()
@@ -1090,6 +1097,77 @@ LIMIT 1
             },
         )
         "###);
+    }
+
+    #[test]
+    fn test_resource_raw_domain() {
+        let json = r#"
+            {
+                "spans": [
+                    {
+                    "timestamp": 1694732408.3145,
+                    "start_timestamp": 1694732407.8367,
+                    "exclusive_time": 477.800131,
+                    "description": "/static/myscript-v1.9.23.js",
+                    "op": "resource.script",
+                    "span_id": "97c0ef9770a02f9d",
+                    "parent_span_id": "9756d8d7b2b364ff",
+                    "trace_id": "77aeb1c16bb544a4a39b8d42944947a3",
+                    "data": {
+                        "http.decoded_response_content_length": 128950,
+                        "http.response_content_length": 36170,
+                        "http.response_transfer_size": 36470,
+                        "resource.render_blocking_status": "blocking",
+                        "server.address": "subdomain.example.com:5688",
+                        "url.same_origin": true,
+                        "url.scheme": "https"
+                    },
+                    "hash": "e2fae740cccd3789"
+                },
+                {
+                    "timestamp": 1694732408.3145,
+                    "start_timestamp": 1694732407.8367,
+                    "exclusive_time": 477.800131,
+                    "description": "/static/myscript-v1.9.24.js",
+                    "op": "resource.script",
+                    "span_id": "97c0ef9770a02f9d",
+                    "parent_span_id": "9756d8d7b2b364ff",
+                    "trace_id": "77aeb1c16bb544a4a39b8d42944947a3",
+                    "data": {
+                        "http.decoded_response_content_length": 128950,
+                        "http.response_content_length": 36170,
+                        "http.response_transfer_size": 36470,
+                        "resource.render_blocking_status": "blocking"
+                    },
+                    "hash": "e2fae740cccd3788"
+                }
+            ]
+            }
+        "#;
+
+        let mut event = Annotated::<Event>::from_json(json)
+            .unwrap()
+            .into_value()
+            .unwrap();
+
+        extract_span_tags(
+            &mut event,
+            &Config {
+                max_tag_value_size: 200,
+            },
+        );
+
+        let span_1 = &event.spans.value().unwrap()[0];
+        let span_2 = &event.spans.value().unwrap()[1];
+
+        let tags_1 = span_1.value().unwrap().sentry_tags.value().unwrap();
+        let tags_2 = span_2.value().unwrap().sentry_tags.value().unwrap();
+
+        assert_eq!(
+            tags_1.get("raw_domain").unwrap().as_str(),
+            Some("https://subdomain.example.com:5688")
+        );
+        assert!(!tags_2.contains_key("raw_domain"));
     }
 
     #[test]